requirements.txt
# requirements.txt
haystack-ai>=2.0.0
faiss-cpu
sentence-transformers
fastapi
uvicorn[standard]
python-dotenv
pandas
orjson